Add : build for mame in progress

+++ scripts/modules/module-mak.ps1
@@ -1,13 +1,12 @@
-Import-Module "..\..\scripts\modules\module-emulators.ps1"
 Import-Module "..\..\scripts\modules\module-iso.ps1"
+Import-Module "..\..\scripts\modules\module-mame-hash.ps1"
 
 function Remove-Project {
   param (
     [Parameter(Mandatory=$true)][String] $ProjectName
   )
-  Write-Host "clean $ProjectName" -ForegroundColor Yellow
+  Write-Host "clean $ProjectName project" -ForegroundColor Yellow
   Write-Host "remove $env:TEMP\neocore\$ProjectName"
-  Stop-Emulators
   if (Test-Path -Path $env:TEMP\neocore\$ProjectName) {
     Get-ChildItem -Path $env:TEMP\neocore\$ProjectName -Recurse -ErrorAction SilentlyContinue | Remove-Item -force -Recurse -ErrorAction SilentlyContinue
   }
@@ -27,18 +26,22 @@ function Set-EnvPath {
 
 function Write-Mame {
   param (
-    [Parameter(Mandatory=$true)][String] $MameBin,
+    [Parameter(Mandatory=$true)][String] $ProjectName,
     [Parameter(Mandatory=$true)][String] $PathMame,
     #[Parameter(Mandatory=$true)][String] $MameArgs,
     [Parameter(Mandatory=$true)][String] $CUEFile,
     [Parameter(Mandatory=$true)][String] $OutputFile
   )
+
+  if ((Test-Path -Path $PathMame) -eq $false) { Write-Host "error - $PathMame not found" -ForegroundColor Red; exit 1 }
+  if ((Test-Path -Path $CUEFile) -eq $false) { Write-Host "error - $CUEFile not founc" -ForegroundColor Red; exit 1 }
+  if ((Test-Path -Path "$PathMame\mame64.exe") -eq $false) { Write-Host "error - mame64.exe is not found in $PathMame" -ForegroundColor Red; exit 1 }
+
   & chdman.exe createcd -i $CUEFile -o $OutputFile --force
-  #& $CHDMANBin createcd -i $CUEFile -o $OutputFile --force
-      <# TODO 
-    %CHDMAN% createcd -i %FILECUE% -o %FILECHD% --force > nul
-    powershell -ExecutionPolicy Bypass -File ..\..\scripts\mame-hash-writer.ps1 %PROJECT% %FILECHD% %MAMEHASH%
-    #>
+  if ((Test-Path -Path $OutputFile) -eq $false) { Write-Host "error - $OutputFile is not found" -ForegroundColor Red; exit 1 }
+
+  Write-MameHash -ProjectName $ProjectName -CHDFile $OutputFile -XMLFile "$PathMame\hash\neocd.xml"
+  & "$PathMame\mame64.exe" -rompath "$PathMame\roms" -hashpath "$PathMame\hash" -cfg_directory $env:TEMP -nvram_directory $env:TEMP -skip_gameinfo neocdz $ProjectName
     <#
     :mame-start-process
       echo starting mame ...
@@ -53,12 +56,19 @@ function Write-CUE {
     [Parameter(Mandatory=$true)][String] $OutputFile,
     [Parameter(Mandatory=$true)][String] $ISOName
   )
-  "CATALOG 0000000000000 " | Out-File -Encoding utf8 -FilePath $OutputFile -Force
-  ('  FILE "{0}" BINARY ' -f $ISOName) | Out-File -Encoding utf8 -FilePath $OutputFile -Append -Force
-  "  TRACK 01 MODE1/2048 " | Out-File -Encoding utf8 -FilePath $OutputFile -Append -Force
-  "  INDEX 01 00:00:00 " | Out-File -Encoding utf8 -FilePath $OutputFile -Append -Force
-  Write-Host "builded CUE is available to $OutputFile" -ForegroundColor Green
-  Write-Host ""
+  "CATALOG 0000000000000 " | Out-File -Encoding ascii -FilePath $OutputFile -Force
+  ('  FILE "{0}" BINARY ' -f $ISOName) | Out-File -Encoding ascii -FilePath $OutputFile -Append -Force
+  "  TRACK 01 MODE1/2048 " | Out-File -Encoding ascii -FilePath $OutputFile -Append -Force
+  "  INDEX 01 00:00:00 " | Out-File -Encoding ascii -FilePath $OutputFile -Append -Force
+  # TODO or not
+  #& unix2dos $OutputFile
+  if ((Test-Path -Path $OutputFile) -eq $true) {
+    Write-Host "builded CUE is available to $OutputFile" -ForegroundColor Green
+    Write-Host ""
+  } else {
+    Write-Host "error - $OutputFile was not generated" -ForegroundColor Red
+    exit 1
+  }
 }
 
 function Write-Program {
@@ -69,13 +79,19 @@ function Write-Program {
     [Parameter(Mandatory=$true)][String] $ProjectName
   )
   Write-Host "compiling program" -ForegroundColor Yellow
+  if ((Test-Path -Path $MakeFile) -eq $false) { Write-Host "error - $MakeFile not found" -ForegroundColor Red; exit 1 }
+
   $env:PROJECT = $ProjectName
   $env:NEODEV = $PathNeoDev
   $env:FILEPRG = $PRGFile
   & make -f $MakeFile
-  # TODO : test outfile
-  Write-Host "builded program is available to $PRGFile" -ForegroundColor Green
-  Write-Host ""
+  if ((Test-Path -Path $PRGFile) -eq $true) {
+    Write-Host "builded program is available to $PRGFile" -ForegroundColor Green
+    Write-Host ""  
+  } else {
+    Write-Host "error - $PRGFile was not generated" -ForegroundColor Red
+    exit 1
+  }
 }
 
 function Write-ISO {
@@ -90,15 +106,15 @@ function Write-ISO {
   if (Test-Path -Path $PathISOBuildFolder) { Remove-Item $PathISOBuildFolder -Recurse -Force }
   if (-Not(Test-Path -Path $PathISOBuildFolder)) { mkdir -Path $PathISOBuildFolder | Out-Null }
   if (-Not(Test-Path -Path $PathCDTemplate)) {
-    Write-Host "error : $PathCDTemplate not found" -ForegroundColor Red
+    Write-Host "error - $PathCDTemplate not found" -ForegroundColor Red
     exit 1
   }
   if (-Not(Test-Path -Path $PRGFile)) {
-    Write-Host "error : $PRGFile not found" -ForegroundColor Red
+    Write-Host "error - $PRGFile not found" -ForegroundColor Red
     exit 1
   }
   if (-Not(Test-Path -Path $SpriteFile)) {
-    Write-Host "error : $SpriteFile not found" -ForegroundColor Red
+    Write-Host "error - $SpriteFile not found" -ForegroundColor Red
     exit 1
   }
   Copy-Item -Path "$PathCDTemplate\*" -Destination $PathISOBuildFolder -Recurse -Force
@@ -106,9 +122,14 @@ function Write-ISO {
   Copy-Item -Path $SpriteFile -Destination "$PathISOBuildFolder\DEMO.SPR" -Force
 
   New-IsoFile -Source $PathISOBuildFolder -Path $OutputFile -Force | Out-Null
-  # TODO : test outfile
-  Write-Host "builded ISO is available to $OutputFile" -ForegroundColor Green
-  Write-Host ""
+  
+  if ((Test-Path -Path $OutputFile) -eq $true) {
+    Write-Host "builded ISO is available to $OutputFile" -ForegroundColor Green
+    Write-Host ""  
+  } else {
+    Write-Host "error - $OutputFile was not generated" -ForegroundColor Red
+    exit 1
+  }
 }
 
 function Write-Sprite {
@@ -118,13 +139,19 @@ function Write-Sprite {
     [Parameter(Mandatory=$true)][String] $XMLFile
   )
   Write-Host "compiling sprites" -ForegroundColor Yellow
-  # TODO : catch error
+  if ((Test-Path -Path $XMLFile) -eq $false) { Write-Host "error - $XMLFile not found" -ForegroundColor Red; exit 1 }
+
   & BuildChar.exe $XMLFile
   & CharSplit.exe char.bin "-$Format" $OutputFile
   Remove-Item -Path char.bin -Force
-  # TODO : test outfile
-  Write-Host "builded sprites is available to $OutputFile.$Format" -ForegroundColor Green
-  Write-Host ""
+  
+  if ((Test-Path -Path "$OutputFile.$Format") -eq $true) {
+    Write-Host "builded sprites is available to $OutputFile.$Format" -ForegroundColor Green
+    Write-Host ""  
+  } else {
+    Write-Host ("error - {0}.{1} was not generated" -f $OutputFile, $Format) -ForegroundColor Red
+    exit 1
+  }
 }
 
 function Write-ZIP {
@@ -134,11 +161,19 @@ function Write-ZIP {
     [Parameter(Mandatory=$true)][String] $OutputFile
   )
   Write-Host "compiling ZIP" -ForegroundColor Yellow
+  if ((Test-Path -Path $ISOFile) -eq $false) { Write-Host "error - $ISOFile not found" -ForegroundColor Red; exit 1 }
+  if ((Test-Path -Path $Path) -eq $false) { Write-Host "error - $Path not found" -ForegroundColor Red; exit 1 }
+
   if ((Test-Path -Path $OutputFile)) { Remove-Item $OutputFile -Force }
   Add-Type -Assembly System.IO.Compression.FileSystem
   $compressionLevel = [System.IO.Compression.CompressionLevel]::Optimal
   [System.IO.Compression.ZipFile]::CreateFromDirectory($Path, $OutputFile, $compressionLevel, $false)
-   # TODO : test outfile
-   Write-Host "builded ZIP is available to $OutputFile" -ForegroundColor Green
-  Write-Host ""
+
+  if ((Test-Path -Path $OutputFile) -eq $true) {
+    Write-Host "builded ZIP is available to $OutputFile" -ForegroundColor Green
+    Write-Host ""  
+  } else {
+    Write-Host "error - $OutputFile was not generated" -ForegroundColor Red
+    exit 1
+  }
 }
